Task error logging in database
This is half of issue #25, the other half being the UI to actually
display the errors.

There is one caveat to the errors that are logged: not everything is
caught in in the scrapy crawler. There isn't a really good way to catch
errors while using ClawlSpider. I think the way to solve this is to add
a middleware that logs errors.