Add crawler stop button. Fixes #14
Note: the "Start" button that runs the crawl/process/query tasks
sequentially is now done using a celery task chain. This way, if one of
them is killed manually, the next will run. Of course, using the
"advanced" buttons, only one is run at a time, even if it's killed.